Transfer CLI
Also add click to standard requirements
